Renamed modules names
