+++ tests/ranfunc/internal/ranfuncparams/ranfuncparams.go
@@ -1,0 +1,20 @@
+package ranfuncparams
+
+const (
+	// HubKubeEnvKey is the hub's kubeconfig env var.
+	HubKubeEnvKey string = "KUBECONFIG"
+	//  SpokeKubeEnvKey is the spoke's kubeconfig env var.
+	SpokeKubeEnvKey string = "KUBECONFIG_SPOKE1"
+	// AcmOperatorName operator name of ACM.
+	AcmOperatorName string = "advanced-cluster-management"
+	// AcmOperatorNamespace ACM's namespace.
+	AcmOperatorNamespace string = "rhacm"
+	// OpenshiftGitops name.
+	OpenshiftGitops string = "openshift-gitops"
+	// OpenshiftGitopsRepoServer ocp git repo server.
+	OpenshiftGitopsRepoServer string = "openshift-gitops-repo-server"
+	// OperatorHubTalmNamespace talm namespace.
+	OperatorHubTalmNamespace string = "topology-aware-lifecycle-manager"
+	// OpenshiftOperatorNamespace ocp namespaces.
+	OpenshiftOperatorNamespace string = "openshift-operators"
+)
