+++ tests/ranfunc/powermanagement/internal/powermanagementhelper/stats.go
