+++ tests/ranfunc/powermanagement/internal/powermanagementparams/powermanagementparams.go
